create `LogzioGrafanaAlertsClient` struct
